chore:cleanup master, address review comments
