+++ src/lib/JSONViewer.js
@@ -1,6 +1,10 @@
 import React from 'react';
 import {isArray, isObject, isNumber, isString, isBoolean} from 'lodash';
 import {CollapseIcon, isNodeCollapsed, toggleNodeCollapsed} from './CollapseIcon';
+import { viewerStyles } from "./helpers/viewerStyles"
+import { getViewerKey } from "./helpers/ViewerKey"
+import { LabelAndValue } from './components/LabelAndValue'
+import { ViewerLabel } from './components/ViewerLabel'
 
 export default class JSONViewer extends React.Component {
   static defaultProps = {
@@ -8,7 +12,7 @@ export default class JSONViewer extends React.Component {
     marginLeftStep: 2, //no of spaces to the left per nested object
     collapsible: false, //whether nodes are collapsible or not
     collapsedNodes: {},
-    styles: {styles} //pass to override styles
+    styles: {viewerStyles} //pass to override styles
   };
 
   constructor(props){
@@ -46,16 +50,16 @@ export default class JSONViewer extends React.Component {
         this.getCollapseIcon(marginLeft, currentKey, parentKeyPath)
       );
     }
-    
+
     if(isNodeCollapsed.call(this, marginLeft, currentKey, marginLeftStep)) return; //this node is collapsed
 
     let prevIsLastSibling = isLastSibling;
     for(let key = 0; key < data.length; key++){
       isLastSibling = key === data.length - 1;
-      elems.push(<br key={getKey('break', key, parentKeyPath, marginLeft)}/>);
+      elems.push(<br key={getViewerKey('break', key, parentKeyPath, marginLeft)}/>);
       this.recursiveParseData(key, parentKeyPath, data, elems, marginLeft + marginLeftStep, isLastSibling);
     }
-    elems.push(<br key={getKey('break', currentKey, parentKeyPath, marginLeft)}/>);
+    elems.push(<br key={getViewerKey('break', currentKey, parentKeyPath, marginLeft)}/>);
     elems.push(this.getLabel(']', 'builtin', marginLeft, prevIsLastSibling, currentKey, parentKeyPath)); //closing array tag
   }
 
@@ -81,12 +85,12 @@ export default class JSONViewer extends React.Component {
     let prevIsLastSibling = isLastSibling;
     keys.forEach(key => {
       isLastSibling = ++count === keys.length;
-      elems.push(<br key={getKey('break', key, parentKeyPath, marginLeft)}/>);
+      elems.push(<br key={getViewerKey('break', key, parentKeyPath, marginLeft)}/>);
       this.recursiveParseData(
         key, parentKeyPath, data, elems, marginLeft + marginLeftStep, isLastSibling);
     });
 
-    elems.push(<br key={getKey('break', currentKey, parentKeyPath, marginLeft)}/>);
+    elems.push(<br key={getViewerKey('break', currentKey, parentKeyPath, marginLeft)}/>);
     elems.push(this.getLabel('}', 'builtin', marginLeft, prevIsLastSibling, currentKey, parentKeyPath)); //closing object tag
   }
 
@@ -100,7 +104,7 @@ export default class JSONViewer extends React.Component {
   }
 
   recursiveParseData(currentKey, parentKeyPath, parent, elems, marginLeft, isLastSibling){
-    let data = parent[currentKey]; 
+    let data = parent[currentKey];
     switch(this.getDataType(data)){
       case "array":
         this.parseArray(currentKey, parentKeyPath, data, parent, elems, marginLeft, isLastSibling);
@@ -128,24 +132,24 @@ export default class JSONViewer extends React.Component {
           this.getLabelAndValue(currentKey, parentKeyPath, data, parent, "builtin", marginLeft, isLastSibling)
         );
     }
-    
+
   }
 
   getCollapseIcon(marginLeft, currentKey, parentKeyPath){
     let {collapsedNodes} = this.state;
     let {collapsible, marginLeftStep, styles} = this.props;
     return (
-      <span key={getKey('collapse_and_remove', currentKey, parentKeyPath, marginLeft)}>
-        <CollapseIcon 
-          collapsedNodes={collapsedNodes} 
-          marginLeft={marginLeft} 
-          collapsible={collapsible} 
+      <span key={getViewerKey('collapse_and_remove', currentKey, parentKeyPath, marginLeft)}>
+        <CollapseIcon
+          collapsedNodes={collapsedNodes}
+          marginLeft={marginLeft}
+          collapsible={collapsible}
           currentKey={currentKey}
           styles={styles}
           isNodeCollapsed={isNodeCollapsed.bind(this, marginLeft, currentKey, marginLeftStep)}
           toggleNodeCollapsed={toggleNodeCollapsed.bind(this, marginLeft, currentKey, marginLeftStep)}
         />
-      
+
       </span>
     )
   }
@@ -157,8 +161,8 @@ export default class JSONViewer extends React.Component {
       return this.getLabel(value, type, marginLeft, isLastSibling, currentKey, parentKeyPath);
     } else {
       return (
-        <LabelAndValue 
-          key={getKey('label_and_value', currentKey, parentKeyPath, marginLeft)}
+        <LabelAndValue
+          key={getViewerKey('label_and_value', currentKey, parentKeyPath, marginLeft)}
           currentKey={currentKey}
           value={value}
           type={type}
@@ -173,10 +177,10 @@ export default class JSONViewer extends React.Component {
   getLabel(value, type, marginLeft, isLastSibling, currentKey, parentKeyPath){
     const {styles} = this.props;
     return (
-      <Label 
-        key={getKey('label', currentKey + value, parentKeyPath, marginLeft)}
+      <ViewerLabel
+        key={getViewerKey('label', currentKey + value, parentKeyPath, marginLeft)}
         value={value}
-        type={type} 
+        type={type}
         marginLeft={marginLeft}
         isLastSibling={isLastSibling}
         styles={styles}/>
@@ -191,93 +195,4 @@ export default class JSONViewer extends React.Component {
   }
 }
 
-const printSpaces = (marginLeft) => {
-  //we would have used css to set a margin left
-  //but that makes the json lose its formatting when copied
-  let spaces = [];
-  for(let x = 0; x < marginLeft; x++){
-    spaces.push(<span key={x}>&nbsp;</span>);
-  }
-  return <span>{spaces}</span>;
-};
 
-const getKey = (prefix, currentKey, parentKeyPath, marginLeft) => {
-  return `${prefix}_${parentKeyPath}_${currentKey}_${marginLeft}`
-}
-
-const Label = (props) => {
-  let {marginLeft, value, type, isLastSibling, styles} = props;
-  let style = styles.text;
-  switch(type){
-    case "number":
-      style = styles.number;
-      if(!isLastSibling) value = value + ",";
-      break;
-    case "boolean":
-      style = styles.builtin;
-      value = value + ""; //coerce boolean to string, seems you cant return booleans in react elements
-      if(!isLastSibling) value = value + ",";
-      break;
-    case "property":
-      style = styles.property;
-      value = "\"" + value + "\":"; //add quotes to string
-      break;
-    case "builtin":
-      style = styles.builtin;
-      value = value + "";
-      if(!isLastSibling) value = value + ",";
-      break;
-    default:
-      style = styles.text;
-      if(isLastSibling)
-        value = "\"" + value + "\"";
-      else
-        value = "\"" + value + "\",";
-  }
-  return (
-    <span style={style}>{printSpaces(marginLeft)}{value}</span>
-  );
-}
-
-const LabelAndValue = (props) => {
-  let {currentKey, marginLeft, type, value, isLastSibling, styles} = props;
-  return (
-    <span key={`label_and_value_${currentKey}`}>
-      <Label 
-        value={currentKey}
-        type="property" 
-        isLastSibling={isLastSibling}
-        marginLeft={marginLeft}
-        styles={styles}/>
-      <Label 
-        value={value}
-        type={type}
-        isLastSibling={isLastSibling}
-        marginLeft={1}
-        styles={styles}/>
-    </span>
-  );
-}
-
-const styles = {
-  root: {
-    margin: 5,
-    fontSize: 14,
-    fontFamily: "monospace"
-  },
-  builtin: {
-    color: "#00f"
-  },
-  text: {
-    color: "#077"
-  },
-  number: {
-    color: "#a0a"
-  },
-  property: {
-    color: "#c00"
-  },
-  collapseIcon: {
-    cursor: "pointer"
-  }
-};
