+++ src/lib/components/AddIcon.js
