+++ src/lib/components/RemoveIcon.js
