--- src/lib/components/jsonEditorComponents/index.js
@@ -1,10 +1,0 @@
-export * from "./styles";
-export * from "./editKey";
-export * from "./Boolean";
-export * from "./Label";
-export * from "./RemoveIcon";
-export * from "./SaveIcon";
-export * from "./AddIcon";
-export * from "./GetKey";
-export * from "./ParentLabel";
-export * from "./Input";
