+++ src/lib/helpers/ViewerKey.js
@@ -1,0 +1,3 @@
+export const getViewerKey = (prefix, currentKey, parentKeyPath, marginLeft) => {
+  return `${prefix}_${parentKeyPath}_${currentKey}_${marginLeft}`;
+};
