+++ src/lib/helpers/viewerStyles.js
@@ -1,0 +1,22 @@
+export const viewerStyles = {
+  root: {
+    margin: 5,
+    fontSize: 14,
+    fontFamily: "monospace",
+  },
+  builtin: {
+    color: "#00f",
+  },
+  text: {
+    color: "#077",
+  },
+  number: {
+    color: "#a0a",
+  },
+  property: {
+    color: "#c00",
+  },
+  collapseIcon: {
+    cursor: "pointer",
+  },
+};
